add some super-crude torque constants
We got these by running 05-gravity-compensation (the new one with the
curses UI for tuning) and steadily increasing the constants from zero
until the arms were able to support themselves against gravity. Note
that these values are crap! The motors seem to flat-out ignore current
commands below a certain magnitude and friction dominates most of the
behavior we're observing. The only use for these results is probably
to be added to other controllers to offset gravity slightly.


Former-commit-id: cb95d6d13968ee93f83e7494ae4443613f33020a