compute the jacobian's secondary goal outside the function
the secondary goal is something the workspace controller should be
doing - the concept behind it is "move the elbow so it's pointed in
this direction", not just "drive joints to zero". As such, it should
be computed outside the strict jacobian function and passed in.

Also, I don't think the nullspace gain is a proportional constant for
a controller (I _think_), so rename it from K_nullspace_p to
nullspace_gain.


Former-commit-id: 1d2a58d3555658b9b027da2c913a47dcf5b946dd